+++ index.html
@@ -233,6 +233,6 @@
   </template>
 
   <!-- Подключение скриптов -->
-  <script src="/js/main.js" type="module"></script>
+  <script src="js/main.js" type="module"></script>
 </body>
 </html>
